requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
tabulate==0.9.0
urllib3==2.2.2
